requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Medical Imaging
monai>=1.3.0

# Training Framework
pytorch-ignite>=0.4.11

# Data Processing
numpy>=1.21.0
scipy>=1.8.0
nibabel>=3.2.0
pandas>=1.3.0

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.62.0
tensorboard>=2.10.0

